separate FormPartsContainer from formHelpers
