+++ src/forms/CurrentIncome.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { Form } from 'semantic-ui-react';
 
 // PROJECT COMPONENTS
-import { FormPartsContainer } from './formHelpers';
+import FormPartsContainer from './FormPartsContainer';
 import { IntervalColumnHeadings } from '../components/headings';
 import { CashFlowRow } from './cashflow';
