+++ src/forms/formHelpers.js
@@ -2,16 +2,9 @@
 import React from 'react';
 import {
   // Generic Form stuff
-  Header,
-  Segment,
   Label,
-  Divider,
   Icon,
 } from 'semantic-ui-react';
-
-// PROJECT COMPONENTS
-import { BigButton } from './inputs';
-
 
 // ========================================
 // GENERIC COMPONENTS
@@ -27,136 +20,6 @@ const ExternalLink = function ({ href, children }) {
       target='_blank'>{children}
     </a>);
 };
-
-
-/** For styling spacing between elements when needed.
-* @returns Component
-*/
-const SpaceHolder = function () {
-  return (<div className = { `space-holder` } />);
-};
-
-
-/** The row containing the big buttons at the bottom of each
-*     form section, such as 'Previous', 'Next', and 'New Client'.
-* 
-* @object buttonProps
-* @property text {string} - Text to show on the button.
-* @property onClick {function} - Optional function to run on button click.
-*
-* @function
-* @param {object} props - One object for each button
-* @property {buttonProps} props.left
-* @property {buttonProps} props.middle
-* @property {buttonProps} props.right
-*
-* @returns Component
-*/
-const FormBottomRow = function({ left, middle, right }) {
-
-  var Left    = <SpaceHolder key = { 'left' } />,
-      Middle  = <SpaceHolder key = { 'middle' } />,
-      Right   = <SpaceHolder key = { 'right' } />;
-
-  if (left) {
-    Left = (
-      <BigButton
-        key = { 'left' }
-        onClick = { left.onClick }>
-        { left.text }
-      </BigButton>
-    );
-  }
-
-  // Considering having a non-button as a label
-  // for the page. Not sure how to preserve the
-  // same style.
-  if (middle) {
-    Middle = (<div key = { 'middle' }>{ middle.text }</div>);
-  }
-
-  if (right) {
-    Right = (
-      <BigButton
-        key = { 'right' }
-        onClick = { right.onClick }>
-        { right.text }
-      </BigButton>
-    );
-  }
-
-  var children = [
-    Left,
-    // Needed for first form section where there's no
-    // 'Prev' button.
-    Middle,
-    Right,
-  ];
-
-  /** @todo Move styles to CSS */
-  return (
-    <div
-      className = { `form-section-bottom-row` }
-      style     = {{ display: `flex`, justifyContent: `space-between` }}>
-      { children }
-    </div>
-  );
-
-}; // End <FormBottomRow>
-
-
-/** Constructor for all the stuff that's supposed to go inside
-* the Form Component. Does not include the `<Form>` Component
-* as a container because it looks like that needs to be unique
-* (the 'CurrentBenefitsStep' gives it `size='massive'`).
-*
-* @function
-* @param {object} props
-* @property {string} props.title - Text to go in the `h1` element.
-* @property {string} props.clarifier - Text to go in the `h3`
-* element, giving some description, instructions, or clarifications.
-* @property {string} props.children - Component(s) to be inserted
-* into the middle - a custom form section containing inputs, etc.
-* @property {Object} props.next - the 'next form section' function
-* @property {Object} props.prev - the 'previous form section' function
-*
-* @returns Component
-*/
-const FormPartsContainer = function({ title, clarifier, children, navData }) {
-  return (
-    <Segment
-      padded='very'
-      className="flex-item flex-column">
-      <Segment
-        basic={ true }
-        className="flex-item">
-        <Header
-          as='h1'
-          color='teal'
-          textAlign='center'>
-          { title }
-        </Header>
-        { !clarifier
-          ? null
-          : (
-            <Header
-              as='h3'
-              textAlign='center'>
-              { clarifier }
-            </Header>
-          )
-        }
-
-        { children }
-
-      </Segment>
-
-      <Divider />
-      <FormBottomRow { ...navData } />
-
-    </Segment>
-  );
-};  // End FormPartsContainer() Component
 
 
 // ========================================
@@ -203,9 +66,7 @@ var AttentionArrow = function () {
 
 /** @todo Separate into different files? */
 export {
-  ExternalLink, SpaceHolder,
-  FormBottomRow,
-  FormPartsContainer,
+  ExternalLink,
   InvalidMessage,
   AttentionArrow,
 };
